+++ src/string/LSD.cpp
@@ -3,10 +3,10 @@
 namespace code047 {
 
 	/// <summary>
-	/// 时间复杂度：O(WN+WR) ~ O(WN) 
-	///			   其中W为字符串长度，R为字符集个数都可视为常数，
-	///			   所以LSD字符串排序算法时间复杂度为线性
-	/// 空间复杂度：O(R+N)~O(N) 使用了aux数组N个元素及count数组R个元素
+	/// 时锟戒复锟接度ｏ拷O(WN+WR) ~ O(WN) 
+	///			   锟斤拷锟斤拷W为锟街凤拷锟斤拷锟斤拷锟饺ｏ拷R为锟街凤拷锟斤拷锟斤拷锟斤拷锟斤拷锟斤拷锟斤拷为锟斤拷锟斤拷锟斤拷
+	///			   锟斤拷锟斤拷LSD锟街凤拷锟斤拷锟斤拷锟斤拷锟姐法时锟戒复锟接讹拷为锟斤拷锟斤拷
+	/// 锟秸间复锟接度ｏ拷O(R+N)~O(N) 使锟斤拷锟斤拷aux锟斤拷锟斤拷N锟斤拷元锟截硷拷count锟斤拷锟斤拷R锟斤拷元锟斤拷
 	/// </summary>
 	/// <param name="strings"></param>
 	/// <param name="w"></param>
@@ -19,22 +19,22 @@ namespace code047 {
 			std::vector<int> count(R + 1, 0);
 
 			// Frequency count
-			for (int j = 0; j < strings.size(); j++) {
+			for (size_t j = 0; j < strings.size(); j++) {
 				count[strings[j][i] + 1]++;
 			}
 
 			// Turn frequency to index in sorted array
-			for (int j = 1; j < R; j++) {
+			for (size_t j = 1; j < R; j++) {
 				count[j] += count[j - 1];
 			}
 
 			// Copy elements to aux with index in count 
-			for (int j = 0; j < strings.size(); j++) {
+			for (size_t j = 0; j < strings.size(); j++) {
 				aux[count[strings[j][i]]++] = strings[j];
 			}
 
 			// Copy back
-			for (int j = 0; j < strings.size(); j++) {
+			for (size_t j = 0; j < strings.size(); j++) {
 				strings[j] = aux[j];
 			}
 		}
